Vadim v0.2.18: -фиксы -очищаем код шаблонов (ощущается будто под ковер все подмел) -именование страниц на хедере -решаблонизация нескольких страниц
(есть готовые формы для редактировании группы, надеюсь завтра вставлю и сделаю формы редактирования уроков)
