+++ it_school/mainpage/templates/personal_cabinet.html
@@ -15,7 +15,7 @@
 
 {% block content %}
 
-<div class="courses-base wr" style="display:flex;">
+<div class="courses-base" style="display:flex;">
     <div class="courses-filter-wrapper">
 
            <select class="form__input" id="technology-filter">
@@ -47,17 +47,21 @@
               </select>
 
     </div>
-    <div class="courses-base wr" style="display:flex;">
+    <div class="courses-base" style="display:flex;height: auto;">
+        <div style="backdrop-filter: invert(20%); width: 100%; height: 100%;justify-content: center;display: flex;">
         <div class="courses-list-wrapper jc-c">
             <div class="courses-list-align">
                 <h2 style="color: black; text-align: center">Имеющиеся курсы:</h2>
                 {% for course in user.courses.all %}
-                <div class="courses-unit-wrapper ib va" data-technology="{{ course.technologies }}" data-difficulty="{{ course.difficulty }}">
+                <div class="courses-unit-wrapper ib va" data-technology="{{ course.technologies }}" data-difficulty="{{ course.difficulty }}" style="background: linear-gradient(1deg, #fffcfa7d, #ededed);border: none;">
+                    <!--background: linear-gradient(1deg, #ffad5a47, #ffad5a);-->
+                    <!--background: linear-gradient(1deg, #ffad5a7d, #ededed);-->
                     <div class="courses-unit-texture">
                         <div class="courses-unit-rating ws-nw va">{{ course.rating }}<img src="/static/dist/img/PNG/icon/rating.png" alt="/static/dist/img/PNG/icon/rating.png" class="icn"></div>
-                        <div class="courses-unit-title"><h2><strong>{{ course.title }}</strong></h2></div>
-                        <div class="courses-unit-a"><a href="{% url 'course_lessons' course.id %}">Перейти к занятиям →</a></div>
+                        <div class="courses-unit-title"><h2 style="background-color: #ffffff00;"><strong>{{ course.title }}</strong></h2></div>
                         <div class="courses-unit-difficulty">{{ course.difficulty }}</div>
+                         <div class="courses-unit-shortdes">{{ course.short_des }}</div>
+                        <div class="courses-unit-a" style="margin-top: 80px;"><button onclick="location.href='{% url 'course_lessons' course.id %}'" class="btn">Перейти к занятиям →</button></div>
                     </div>
                 </div>
                 {% empty %}
@@ -66,7 +70,8 @@
             </div>
         </div>
     </div>
-<div style="background-color: #ddddd7; margin-bottom: 0; padding: 20px;">
+    </div>
+<div class="courses-filter-wrapper">
 
         <h3>Информация о пользователе:</h3>
         <p>Имя пользователя: {{ user.username }}</p>
@@ -74,28 +79,13 @@
         <p>Фамилия: {{ user.last_name }}</p>
         <p>Email: {{ user.email }}</p>
         <p>Номер телефона: {{ user.phone_number }}</p>
-        <a id="edit-profile" class="edit-profile-button">Редактировать профиль</a>
-        <h4 style="color: black;">Кошелек:</h4>
-        <p style="color: black;">Баланс: {{ user.wallet }}</p>
-        <a id="edit-wallet" class="edit-profile-button">Пополнить кошелек</a>
+        <button id="edit-profile" class="btn">Редактировать профиль</button>
+        <h4>Кошелек:</h4>
+        <p>Баланс: {{ user.wallet }}</p>
+        <button id="edit-wallet" class="btn">Пополнить кошелек</button>
 
 </div>
 </div>
-<style>
-.edit-profile-button {
-  display: inline-block;
-  padding: 10px 20px;
-  background-color: rgba(255, 165, 0, 0.5);
-  color: black;
-  text-decoration: none;
-  border-radius: 4px;
-  transition: background-color 0.3s ease;
-}
-
-.edit-profile-button:hover {
-  background-color: #ddd;
-}
-</style>
 {% include "edit_profile.html" %}
 {% include "wallet_replenishment.html" %}
 {% endblock %}
