+++ it_school/management/urls.py
@@ -12,5 +12,6 @@ urlpatterns = [
     path('lessons/', views.LessonView.as_view(), name='lesson_view'),
     path('courses/<int:pk>/remove_participant/', views.remove_participant, name='remove_participant'),
     path('courses/<int:pk>/add_participant/', views.add_participant, name='add_participant'),
+    path('lessons/', views.lesson_list, name='lesson_list'),
     # path('lesson/update/<int:pk>/', views.lesson_update, name='lesson_update')
 ]
